utygguyv

+++ Gppbox/Dice.hpp
@@ -34,7 +34,7 @@ public:
 		return distrib(gen);
 	}
 
-	static int randFloat(float min, float max)
+	static float randFloat(float min, float max)
 	{
 		std::random_device rd;
 		std::mt19937 gen(rd());
